On-line discussion/collaboration tool
Klabr is an on-line tool to promote interactive discussions for
development projects and other purposes. It uses Sharepoint Server as a
backend database, but it could easily be adapted to other database
formats. The current client is a Windows application, but this could
also be adapted to using a web interface.